Make useFeatures to be always false
