+++ carbon-feature-plugin/src/main/java/org/wso2/maven/p2/utils/P2Constants.java
@@ -67,7 +67,7 @@ public class P2Constants {
             public static final String ID = "carbon.product";
             public static final String APPLICATION = "carbon.application";
             public static final String RUNTIME_FEATURE = "org.wso2.carbon.runtime";
-            public static final Boolean USE_FEATURES = true;
+            public static final Boolean USE_FEATURES = false;
             public static final Boolean INCLUDE_LAUNCHERS = true;
             public static final String CONFIG_INI_USE = "default";
             public static final String LAUNCHER_NAME = "eclipse";
